admin

+++ klpd/src/main/java/com/klpdapp/klpd/Repository/CartRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.klpdapp.klpd.model.Cart;
 import com.klpdapp.klpd.model.Product;
 
-public interface CartRepo extends JpaRepository <Cart, Long> {
+public interface CartRepo extends JpaRepository <Cart, Integer> {
 
     Cart findByProduct(Product product);
